modif front

+++ app/javascript/controllers/appointment_historic_controller.js
@@ -12,22 +12,26 @@ export default class extends Controller {
     window.location.href = '/desks/' + deskId;
   }
 
-  remove(event){
+  remove(event) {
     console.log("remove")
-    console.log(event.currentTarget.dataset.id)
-    const appointmentId = event.currentTarget.dataset.id;
-    const csrfToken = document.querySelector('meta[name="csrf-token"]').getAttribute('content');
-    const url = '/appointments/' + appointmentId;
-    console.log(url)
-    fetch('/appointments/' + appointmentId, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-        "X-CSRF-Token": csrfToken
-      }
-    })
-    .then(response => {
-        window.location.reload();
-      });
+    console.log(event.currentTarget.dataset.id);
+    if (window.confirm("supprimes tu le rdv?")) {
+      const appointmentId = event.currentTarget.dataset.id;
+      const csrfToken = document.querySelector('meta[name="csrf-token"]').getAttribute('content');
+      const url = '/appointments/' + appointmentId;
+      console.log(url)
+      fetch('/appointments/' + appointmentId, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+          "X-CSRF-Token": csrfToken
+        }
+      })
+        .then(response => {
+          window.location.reload();
+        });
     }
+
+
+  }
 }
